Heretic: always do a full recalc of SNDCURVE on changing SFX volume
Non-full process only updates zero (NULL) level, which is used for local sounds emitted by game menu and player. As a result, the rest of the game world sounds are not properly updated during demo playback.

Full process takes about 100-120 MICROseconds (μs) and can be stricted to "gamestate == GS_LEVEL", but then the non-full process is simply left unused, so it is safe to remove it.