Hexen: always show keys and armor on the automap with the status bar
That means it's the logic of the vanilla automap, and it's better to stick to it. DSDA also always shows the keys and armor bar regardless of the overlay mode.

Thanks to @electricbrass for pointing it out!

This also fixes the lack of update/redraw for the armor icons when the armor level decreases while the automap is active. In case keys are taken away, their icons will also be redrawn, though I'm not sure such a situation ever occurs during a level rather than when switching hubs.

Fixes #191.

Co-Authored-By: Mia McMahill <84749759+electricbrass@users.noreply.github.com>