Heretic: add support for H+H line specials
This commit adds support for H+H-specific linedef effects, both inherited from Boom and custom-made for the KEX engine. The latter were written manually, almost from scratch, and while they may not be technically identical, they behave nearly the same. Boom's "transfer lights" effects are not implemented, but they are extremely rare and not critical.

This should fix the remaining softlocks on the levels.